+++ gegelatilib/src/learn/adversarialLearningAgent.cpp
@@ -154,7 +154,7 @@ std::shared_ptr<Learn::EvaluationResult> Learn::AdversarialLearningAgent::
         while (!ale.isTerminal() &&
                nbActions < this->params.maxNbActionsPerEval) {
             // Get the actions
-            std::vector<uint64_t> actionsID =
+            std::vector<double> actionsID =
                 tee.executeFromRoot(*((const TPG::TPGTeam*)*rootsIterator),
                                     le.getInitActions(),
                                     params.nbEdgesActivable)
