+++ gegelatilib/src/learn/learningAgent.cpp
@@ -156,7 +156,7 @@ std::shared_ptr<Learn::EvaluationResult> Learn::LearningAgent::evaluateJob(
         while (!le.isTerminal() &&
                nbActions < this->params.maxNbActionsPerEval) {
             // Get the actions
-            std::vector<uint64_t> actionsID =
+            std::vector<double> actionsID =
                 tee.executeFromRoot(*root, le.getInitActions(),
                                     this->params.nbEdgesActivable)
                     .second;
