+++ gegelatilib/src/tpg/instrumented/tpgExecutionEngineInstrumented.cpp
@@ -64,7 +64,7 @@ std::vector<const TPG::TPGEdge*> TPG::TPGExecutionEngineInstrumented::
     return winningEdges;
 }
 
-std::pair<std::vector<const TPG::TPGVertex*>, std::vector<uint64_t>> TPG::
+std::pair<std::vector<const TPG::TPGVertex*>, std::vector<double>> TPG::
     TPGExecutionEngineInstrumented::executeFromRoot(
         const TPG::TPGVertex& root, const std::vector<uint64_t>& initActions,
         uint64_t nbEdgesActivated)
